refactor(viewer): drop DuckDB + SQL editor, rename to Filter tab
Single end-user-friendly tab now: chip-based Filter, in-memory
evaluator, single Run button. The DuckDB engine, SQL editor, schema
browser, templates, error rewriter, and saved-SQL-queries module are
gone — Path-B handles the production UX comfortably (4M-entity
runs via index prefilter + cheap-first ordering + chunked yielding)
so the dual-engine split was carrying complexity for no gain.

Removed
  - apps/viewer/src/components/viewer/SearchModal.sql.tsx
  - apps/viewer/src/components/viewer/SearchModal.sql.builder.tsx (renamed)
  - apps/viewer/src/lib/search/sql-state.ts (DuckDB lifecycle)
  - apps/viewer/src/lib/search/sql-templates.ts + .test.ts
  - apps/viewer/src/lib/search/sql-error-rewriter.ts + .test.ts
  - apps/viewer/src/lib/search/sql-builder.ts + .test.ts +
    sql-builder.filter-rules.test.ts (the FilterRule[] → SQL emitter is
    no longer needed; the evaluator consumes the same shape directly)
  - apps/viewer/src/lib/search/saved-queries.ts + .test.ts (was SQL
    strings; saved-filters.ts already covers the FilterRule[] shape)

Added / renamed
  - SearchModal.filter.tsx — orchestrator: run lifecycle (cancel +
    progress + limit-hit badge), Tier-1 candidate folding, federated
    evaluator call, virtualised result table, CSV/JSON export, row
    click → multi-model selection. Replaces SearchModal.sql.tsx.
  - SearchModal.filter.builder.tsx — chip palette; same per-kind
    editors (storey/ifcType/predefined/name/property/quantity) minus
    the old "Generated SQL" preview, "Open in Editor", "Run SQL",
    and "Fast Run only" badge. Single component owns rule editing.
  - common-ifc-types.ts — extracted from the deleted sql-builder.ts;
    fallback type list when entityIndex.byType isn't populated yet.

Slice migration (searchSlice.ts)
  Removed: searchSqlMode, searchSqlQuery, setSearchSqlMode,
  setSearchSqlQuery, SearchSqlMode type. Renamed:
  searchSqlResult/Running/Error → searchFilterResult/Running/Error
  (and their setters). The result/error notebook-pairing rule
  (success clears prior error, error preserves prior result) is
  unchanged and still tested.

UI polish
  - Tab labels: "Search" + "Filter" (was "SQL"). SlidersHorizontal
    icon on the Filter tab.
  - Empty rule list: "Add a rule to start filtering — pick by storey,
    IFC type, name, property, or quantity."
  - Empty results: "0 matches — broaden the rules, lower the limit,
    or try OR."
  - Run button: single primary action (no Fast Run / Run SQL split).
  - Run summary line: "N rules · AND · limit 500" — replaces the
    Generated SQL preview pane.
  - Multi-model banner: "Filtering across all N loaded models. Click
    any row to select that element in the right model." — replaces
    the old SQL Editor / Fast Run split banner.
  - Cancel button replaces Run during in-flight evaluation; modal
    unmount aborts whatever's running.
  - Inline search bridge keeps the "Add 'foo' as rule" affordance
    when the toolbar query is non-empty.

Test status
  - 377/377 src/lib/search + src/store/slices.
  - Full viewer suite 637/638 (1 skipped, 0 fail).
  - 26/26 ESM smoke.
  - Production build green (turbo, ~22s).
  - Type-check clean on every touched file.
