--- apps/viewer/src/lib/search/sql-builder.filter-rules.test.ts
@@ -1,185 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-import { describe, it } from 'node:test';
-import assert from 'node:assert';
-import { generateSqlFromFilterRules, isSqlSupported } from './sql-builder.js';
-import { Rule } from './filter-rules.js';
-
-describe('generateSqlFromFilterRules — empty input', () => {
-  it('returns "" when no rules provided', () => {
-    assert.strictEqual(generateSqlFromFilterRules([]), '');
-  });
-});
-
-describe('generateSqlFromFilterRules — set rules', () => {
-  it('emits IN / NOT IN for ifcType', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType(['IfcWall', 'IfcDoor'])]);
-    assert.match(sql, /e\.type IN \('IfcWall', 'IfcDoor'\)/);
-
-    const negated = generateSqlFromFilterRules([Rule.ifcType(['IfcWall'], 'notIn')]);
-    assert.match(negated, /e\.type NOT IN \('IfcWall'\)/);
-  });
-
-  it('emits FALSE for an empty IN list (vacuous match)', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType([])]);
-    assert.match(sql, /\(FALSE\)/);
-  });
-
-  it('emits TRUE for an empty NOT IN list', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType([], 'notIn')]);
-    assert.match(sql, /\(TRUE\)/);
-  });
-});
-
-describe('generateSqlFromFilterRules — name rule', () => {
-  it('eq is case-insensitive', () => {
-    const sql = generateSqlFromFilterRules([Rule.name('eq', 'Wall-A')]);
-    assert.match(sql, /LOWER\(e\.name\) = LOWER\('Wall-A'\)/);
-  });
-
-  it('contains escapes LIKE metacharacters', () => {
-    const sql = generateSqlFromFilterRules([Rule.name('contains', '50%')]);
-    // "50%" → "50\%" (escape the % so it isn't a wildcard).
-    assert.ok(sql.includes("LIKE '%50\\%%' ESCAPE '\\'"), sql);
-  });
-
-  it('startsWith uses prefix LIKE', () => {
-    const sql = generateSqlFromFilterRules([Rule.name('startsWith', 'EXT')]);
-    assert.ok(sql.includes("LIKE 'ext%' ESCAPE '\\'"), sql);
-  });
-});
-
-describe('generateSqlFromFilterRules — property rule', () => {
-  it('isSet emits EXISTS', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.property('Pset_WallCommon', 'IsExternal', 'isSet', ''),
-    ]);
-    assert.match(sql, /EXISTS \(SELECT 1 FROM properties p/);
-    assert.match(sql, /LOWER\(p\.pset_name\) = LOWER\('Pset_WallCommon'\)/);
-    assert.match(sql, /LOWER\(p\.prop_name\) = LOWER\('IsExternal'\)/);
-  });
-
-  it('isNotSet emits NOT EXISTS', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.property('Pset_WallCommon', 'IsExternal', 'isNotSet', ''),
-    ]);
-    assert.match(sql, /NOT EXISTS \(SELECT 1 FROM properties p/);
-  });
-
-  it('eq compares the COALESCE of typed columns', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.property('Pset_X', 'Status', 'eq', 'Approved'),
-    ]);
-    assert.match(sql, /COALESCE\(p\.value_string,/);
-    assert.match(sql, /= LOWER\('Approved'\)/);
-  });
-
-  it('numeric op parses the rule value as a real and uses value_real', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.property('Pset_WallCommon', 'ThermalTransmittance', 'lt', '0.3'),
-    ]);
-    assert.match(sql, /COALESCE\(p\.value_real, CAST\(p\.value_int AS DOUBLE\)\) < 0\.3/);
-  });
-});
-
-describe('generateSqlFromFilterRules — quantity rule', () => {
-  it('emits EXISTS over quantities with numeric op', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.quantity('Qto_WallBaseQuantities', 'NetSideArea', 'gte', 5),
-    ]);
-    assert.match(sql, /EXISTS \(SELECT 1 FROM quantities q/);
-    assert.match(sql, /q\.value >= 5/);
-  });
-});
-
-describe('generateSqlFromFilterRules — combinator', () => {
-  it('AND joins predicates with AND glue', () => {
-    const sql = generateSqlFromFilterRules(
-      [Rule.ifcType(['IfcWall']), Rule.name('contains', 'EXT')],
-      { combinator: 'AND' },
-    );
-    assert.match(sql, /AND \(LOWER\(e\.name\)/);
-  });
-
-  it('OR joins predicates with OR glue', () => {
-    const sql = generateSqlFromFilterRules(
-      [Rule.ifcType(['IfcWall']), Rule.name('contains', 'EXT')],
-      { combinator: 'OR' },
-    );
-    assert.match(sql, /OR \(LOWER\(e\.name\)/);
-  });
-});
-
-describe('generateSqlFromFilterRules — limit & ordering', () => {
-  it('appends LIMIT when > 0', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType(['IfcWall'])], { limit: 50 });
-    assert.match(sql, /LIMIT 50;$/);
-  });
-
-  it('omits LIMIT when 0', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType(['IfcWall'])], { limit: 0 });
-    assert.doesNotMatch(sql, /LIMIT/);
-  });
-
-  it('always orders by name for deterministic output', () => {
-    const sql = generateSqlFromFilterRules([Rule.ifcType(['IfcWall'])]);
-    assert.match(sql, /ORDER BY e\.name/);
-  });
-});
-
-describe('generateSqlFromFilterRules — storey rule (Codex P1 fix)', () => {
-  it('joins via entities.contained_in_storey + IfcBuildingStorey lookup, not the missing entity_storeys view', () => {
-    const sql = generateSqlFromFilterRules([Rule.storey(['Level 1', 'Level 2'])]);
-    // Self-join through the entities table — no reference to a
-    // non-existent `entity_storeys` view.
-    assert.match(sql, /e\.contained_in_storey IN/);
-    assert.match(sql, /FROM entities s/);
-    assert.match(sql, /s\.type = 'IfcBuildingStorey'/);
-    assert.match(sql, /s\.name IN \('Level 1', 'Level 2'\)/);
-    assert.doesNotMatch(sql, /entity_storeys/);
-  });
-
-  it('NOT IN op flips the outer comparator without changing the inner subquery', () => {
-    const sql = generateSqlFromFilterRules([Rule.storey(['Level 1'], 'notIn')]);
-    assert.match(sql, /e\.contained_in_storey NOT IN/);
-    assert.match(sql, /s\.type = 'IfcBuildingStorey'/);
-  });
-});
-
-describe('generateSqlFromFilterRules — predefinedType rule (Codex P1 fix)', () => {
-  it('isSqlSupported returns false for predefinedType rules', () => {
-    assert.strictEqual(isSqlSupported(Rule.predefinedType(['SOLIDWALL'])), false);
-    assert.strictEqual(isSqlSupported(Rule.ifcType(['IfcWall'])), true);
-  });
-
-  it('drops predefinedType rules from the WHERE clause and surfaces a warning comment', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.ifcType(['IfcWall']),
-      Rule.predefinedType(['SOLIDWALL']),
-    ]);
-    // The WHERE clause holds only the IfcType predicate.
-    assert.match(sql, /e\.type IN \('IfcWall'\)/);
-    assert.doesNotMatch(sql, /predefined_type/);
-    // Header comment lists how many rules were skipped + their kinds.
-    assert.match(sql, /WARNING: 1 rule\(s\) skipped \(predefinedType\)/);
-  });
-
-  it('emits no WHERE when every rule is unsupported', () => {
-    const sql = generateSqlFromFilterRules([
-      Rule.predefinedType(['A']),
-      Rule.predefinedType(['B']),
-    ]);
-    // Two rules dropped → header warns, body has no WHERE.
-    assert.match(sql, /WARNING: 2 rule\(s\) skipped/);
-    assert.doesNotMatch(sql, /\bWHERE\b/);
-  });
-});
-
-describe('generateSqlFromFilterRules — escaping', () => {
-  it('doubles single quotes in user input', () => {
-    const sql = generateSqlFromFilterRules([Rule.name('eq', "O'Brien")]);
-    assert.match(sql, /LOWER\('O''Brien'\)/);
-  });
-});
