+++ apps/viewer/src/store/index.ts
@@ -398,11 +398,9 @@ const createViewerStore = () => create<ViewerState>()((...args) => ({
       searchModalOpen: false,
       searchFieldFilter: 'all',
       searchModelFilter: null,
-      searchSqlQuery: '',
-      searchSqlResult: null,
-      searchSqlRunning: false,
-      searchSqlError: null,
-      searchSqlMode: 'builder',
+      searchFilterResult: null,
+      searchFilterRunning: false,
+      searchFilterError: null,
       searchFilter: { rules: [], combinator: 'AND', limit: 500 },
       searchFilterSchema: new Map(),
     });
